Refactor: mutate form value to modify 'start_at' and account_number
