+++ src/pages/ScheduledPayment/CreateScheduledPayment.tsx
@@ -5,14 +5,13 @@ import * as Yup from 'yup';
 import BulkImport from '../../components/BulkImport';
 import SearchUserInline from '../../components/SearchUserInline';
 import InlineNotification from '../../components/InlineNotification';
-import createSchedulePaymentTemplate from '../../assets/bulk-import-templates/create_scheduled_payment_template.xlsx';
+import createSchedulePaymentTemplate from '../../assets/bulk-import-templates/create_scheduled-payment_template.xlsx';
 
 const Create = () => {
   const [scheduledPaymentID, setScheduledPaymentID] = useState('');
   const [successMessage, setSuccessMessage] = useState('');
   const [errorMessage, setErrorMessage] = useState('');
   const [isLoading, setLoading] = useState(false);
-  const [userNotFound, setUserNotFound] = useState(false);
   const [selectedUser, setSelectedUser] = useState('');
   const [inputFormType, setInputFormType] = useState('one'); // one or multiple
 
@@ -26,7 +25,7 @@ const Create = () => {
       amount: '',
       cause: '',
       recurring: '',
-      start_at: new Date().getTime() / 1000,
+      start_at: '',
     },
 
     validationSchema: Yup.object({
@@ -49,10 +48,12 @@ const Create = () => {
       setSuccessMessage('');
       setErrorMessage('');
 
-      values.start_at = new Date(values.start_at).getTime() / 1000;
-      values.account_number = selectedUser;
       authAxios
-        .post('/scheduled-payment/create', values)
+        .post('/scheduled-payment/create', {
+          ...values,
+          account_number: selectedUser,
+          start_at: values.start_at ? new Date(values.start_at).getTime() / 1000 : '',
+        })
         .then((res) => {
           setScheduledPaymentID(res.data.id);
           setLoading(false);
@@ -145,8 +146,10 @@ const Create = () => {
 
           <SearchUserInline
             query={formik.values.account_number}
-            onSelecteUser={(value) => setSelectedUser(value)}
-            onUserNotFound={(value) => setUserNotFound(value)}
+            onSelecteUser={(value) => {
+              setSelectedUser(value);
+              formik.setFieldValue('account_number', value);
+            }}
           />
 
           <div className="grid md:grid-cols-2 md:gap-6">
@@ -244,7 +247,7 @@ const Create = () => {
 
           <button
             type="submit"
-            disabled={userNotFound || !selectedUser || isLoading}
+            disabled={!selectedUser || isLoading}
             className="text-white bg-violet-600 hover:bg-violet-700 focus:ring-4 focus:outline-none focus:ring-violet-300 font-medium rounded-lg text-sm w-full sm:w-auto px-5 py-2.5 text-center"
           >
             {isLoading ? 'Please wait...' : 'Schedule Payment'}
